# packages/backend/openapi.yaml
openapi: 3.1.0
info:
  title: ProAppStore Platform API
  version: 1.0.0
  description: 'REST API for ProAppStore (PAS) — the Pro app marketplace platform.


    Powers identity, app provisioning + deploy, the per-app backend services (KV, counters, storage, rooms,
    secrets, AI, email/SMS, maps, push), the storefront + submissions review flow, the Agent Teams build
    loop, billing, and creator payouts.


    Auth: send `Authorization: Bearer <PAS session JWT>` (HS256, from `/v1/auth/exchange` or an OAuth
    flow). A handful of service-to-service routes use the `X-Internal-Token` header instead. Public routes
    need no auth.


    This spec is generated from the backend route table; request/response bodies are typed for the core
    resources and generic (object) elsewhere.'
servers:
- url: https://api.proappstore.online
  description: Production
security:
- bearerAuth: []
...
paths:
...
  /v1/apps/{appId}/actions/{name}:
    post:
      tags:
      - Actions
      summary: Create actions
      operationId: post_v1_apps_appId_actions_name
      description: "Run a registered action. Auth: a PAS session JWT, OR a personal app token `Authorization: Bearer pas_at_…` minted for THIS app (#154; roles fixed to `user`, read tokens may call query actions only, action-scoped tokens only their actions). Public actions (requires_auth false) need no auth. A `verify` action (#148) runs its scoped SELECT, hands the rows to the platform verifier named in its manifest (e.g. `chess.replay`), and — when the verifier completes — runs its write statements in one transaction with the verdict bound as `:__verify_<output>`; it answers `{ ok, verifier, output, error?, writes? }` (200 with `ok: false` and no write when the input could not be verified)."
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - bearerAuth: []
      parameters:
      - name: appId
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
components:
  securitySchemes:
    githubOidc:
      type: http
      scheme: bearer
      bearerFormat: GitHub Actions OIDC JWT
      description: 'A GitHub Actions OIDC token (permissions: id-token: write; audience https://api.proappstore.online). Accepted by the keyless deploy routes and POST /v1/auth/exchange/oidc.'
    appToken:
      type: http
      scheme: bearer
      bearerFormat: pas_at_…
      description: 'Personal app token (#154): per-user, per-app, revocable; accepted by POST /v1/apps/{appId}/actions/{name} only. Never by MCP or any other route.'
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: PAS session JWT (HS256). uid/login/roles/appRoles claims.
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Shared secret for service-to-service calls.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
      - error
...